requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
cryptography==43.0.1
dnspython==2.6.1
docker==7.1.0
email_validator==2.2.0
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==28.1.0
fastapi==0.112.2
fastapi-cli==0.0.5
freezegun==1.5.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msal==1.31.0
multidict==6.1.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psutil==5.9.8
psycopg==3.2.1
psycopg-binary==3.2.1
pwdlib==0.2.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
rich==13.8.0
ruff==0.6.3
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.33
starlette==0.38.4
taskipy==1.13.0
testcontainers==4.8.1
tomli==2.0.1
twilio==9.3.0
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.23.2
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
wheel==0.44.0
wrapt==1.16.0
yarl==1.11.1
